Bug 1697346 - Remove snapping of inflated surface rects. r=aosmond,gfx-reviewers
When an offscreen surface establishes a raster root, this code
was causing incorrect snapping / rounded (even on non-inflated
surfaces), resulting in test failures in some cases.

In theory, this should not be necessary, since scroll offsets are
snapped, and primitives are already snapped during scene building.
Additionally, the picture surface allocation code expects surfaces
with fractional offsets and handles this case (by rounding out the
allocation size, and creating a UV set that samples from the subpixel
offsets of the surface).

In practice, there may be content that relies on this which isn't
tested by CI, so let's land this as a separate patch and see if it
causes any real-world content regressions, before landing the
changes that rely on this.

Differential Revision: https://phabricator.services.mozilla.com/D107768

UltraBlame original commit: 5c91811dabc640ec20d22b96f8d0f683a16efa7f