Bug 1697483: Fix GetInlinedArgument regalloc on x64 r=jandem
The testcase tries to encourage the result of the GetInlinedArgument to be assigned to the same physical register as one of the arguments by tying up as many registers as possible with live values. It fails on x64 without the patch.

Differential Revision: https://phabricator.services.mozilla.com/D107957

UltraBlame original commit: 2ea5fbf4276675f1cea503a0f302e62752d3b538