Bug 1696460 - Allow multiple runtimes to run at the same time. r=aklotz
For tests, we want to be able to run multiple runtimes at the same time (each
runtime gets its own process).

This mostly work, except our process allocator does not currently check that it
owns a certain process and it will happily try to take control of a different
runtime's process (and fail).

To make this work we tag each content process with the parent process' UUID so
that we can check whether the process that we try to create is free or not.

This strategy is pretty rudimental and we can iteratively improve on it if we
decide to use multiple runtimes on actual apps (and not just tests).

Differential Revision: https://phabricator.services.mozilla.com/D108037

UltraBlame original commit: 9c4199568bbc4f6f1a7c439149c744156892c788