Bug 1697816: Fix moveValue when src == typeReg r=jandem
This is very similar to bug 1697483, in that it happens because we are reusing the source of a moveValue as part of the destination. I glanced at the 32-bit code when I fixed that, but missed this problem.

We should not set the tag register until after we copy the src into the payloadReg, because that breaks when src and typeReg are assigned the same register. Swapping the order fixes the problem.

I fixed MIPS while I was at it. ARM64 doesn't need a fix, [because it can set the tag without a scratch register](https://searchfox.org/mozilla-central/rev/f6ffb71dca9eb491e85aa95042380b2602008b00/js/src/jit/arm64/MacroAssembler-arm64.cpp#51-53).

I'm not adding a new testcase here, because I already gave it my best shot in bug 1697483 and it doesn't catch this problem. Specific register allocation patterns are hard to test. It does seem like the fuzzer catches these cases quickly, though.

Differential Revision: https://phabricator.services.mozilla.com/D108084

UltraBlame original commit: 6b87ba86385f42b49f72f7a42e593aa06bd76899