Bug 1697023 - Add end margin/padding back to urlbar tips and their buttons. r=harry
This fixes a couple of problems introduced in bug 1692526 with tips that don't
have help buttons:

The main button focus ring is cut off on the right edge as shown in the
bug. This patch fixes that by adding:

    .urlbarView-tip-button {
      margin-inline-end: 4px;
    }

The main button is too close to the edge of the row, fixed by:

    .urlbarView-row[type=tip] {
      padding-inline-end: calc(urlbarViewIconMarginEnd);
    }

`urlbarViewIconMarginEnd` was removed recently in bug 1695022 (D107568), I'm
guessing because it was only used in one place, but I need it here so I've added
it back. It corresponds to the old `urlbarViewItemInlinePadding` variable that
we used for tips before I incorrectly stopped using it in bug 1692526 [1].

I didn't take tips without help buttons into account when I fixed bug 1692526.
Tips with help buttons look fine currently. Their help buttons are closer to the
edge of the row than they used to be, but that was intentional because I think
it looks better since the help button doesn't have a background like the main
button does. This patch reverts that because it's not possible to keep without
adding a new class or attribute on tips with/without help buttons, and I don't
think it's worth it since we can just revert back to the status quo we had for a
long time, which looks fine too.

[1] https://hg.mozilla.org/mozilla-central/diff/87ab179e18da832a091aa4150ba0e6589baa43ec/browser/themes/shared/urlbarView.inc.css#l1.39

Differential Revision: https://phabricator.services.mozilla.com/D108282

UltraBlame original commit: e6af362362a835be30bd356f17c3ecefb6fa5a49