Bug 1698565 - [devtools] Fix browser_toolbox_backward_forward_navigation.js. r=ochameau.
The test wasn't doing a target switch, which it was supposed to do, so this patch
fixes that.
The other issue was that in the case we're navigating back and forth immediately,
we might get into cases where the `goForward` command was ignored. The fix here
is to use `waitForLocationChange` which does wait just enough so the navigation
was taken into account.

Differential Revision: https://phabricator.services.mozilla.com/D108477

UltraBlame original commit: 363675068fc2690e24b61a79c002c0d938a657e1