Bug 1699000: Use GetCrossDocParentFrameInProcess() in TextOverflow.cpp. r=tnikkel
This patch doesn't change behavior; GetCrossDocParentFrameInProcess() is just a
wrapper for GetCrossDocParentFrame(), which is what we were calling before.

The "InProcess" version of this API (which we're migrating to in this patch) is
used to annotate GetCrossDocParentFrame() callsites that have been vetted as
being OK with the fact that this API returns null at the boundary of a
cross-origin iframe, if fission is enabled.

In this patch, the one call that I'm migrating is inside of
IsFrameDescendantOfAny(), which is a helper-function called by
TextOverflow::PruneDisplayListContents(), which is a function that we use to
traverse a display list and hide content that's been hidden via
"text-overflow:ellipsis" or "-webkit-line-clamp".

We're OK to migrate this call to the "InProcess"-flavored API, because it's
just looking back up the frame tree from the display items that we can reach in
our traversal, which are trivially all part of the same process.

Differential Revision: https://phabricator.services.mozilla.com/D108710

UltraBlame original commit: af9cc260ae23e2185f0bab377f3e51722a37e2e0