Bug 1698662 - Remove misleading comment. r=harry
It's fine if the icon loader outlives the nsMenuItemIconX object.
In the nsMenuItemIconX destructor we call mIconLoader->Destroy() so the
icon loader will not have a dangling reference to nsMenuItemIconX.
I fixed this in bug 1691861 and forgot to remove this comment.

Depends on D108542

Differential Revision: https://phabricator.services.mozilla.com/D108543

UltraBlame original commit: dc1c96ca5060fddbd856bab7b7b51c64b50e8c1e