Bug 1687685 - Use an empty string as default value for request_integrity rather than NULL; r=dom-storage-reviewers,asuth
We have this issue because:
- We didn't check the returning nsresult after executing "VACUUM".
- The mismatch settings between upgrade code for the column request_integrity
and new table code for the column. (one is allowed to be NULL while the other
isn't)
- We rewrite the entries schema.

What we need to fix here are:
- The 21- schema databases. -> Fixing migration code from 21 to 22 should help.
- The 22+ schema databases. -> Requring a new upgrade code.

An alternative option is that maybe we can allow request_integrity to be NULL on
table entries.

Differential Revision: https://phabricator.services.mozilla.com/D107237

UltraBlame original commit: bdfa69ff97f2c0768359fc5b8bca6086777f0c92