Bug 1699682 - Consistently initialize UI state to hide focusrings by default. r=mhowell
UIS_INITIALIZE does something like setting the flag if the last input event was
a mouse event or clearing it if it was a keyboard event. Unfortunately, if this
is initialized to always show focus rings we start always showing outlines for
all content, all the time, which is both undesired and confusing.

It's also not clear from the docs _which_ event it looks at specially at
startup, but anyhow the result we get is clearly flaky, from my testing.

Explicitly clear the flag. It's not clear to me if other applications can cause
the state to change... but otherwise maybe we can just remove the code dealing
with these flags?

Differential Revision: https://phabricator.services.mozilla.com/D109086

UltraBlame original commit: 2dc3348b4132571a5ccb1552725bc848c1543d27