Bug 1691586 - [devtools] Enable browser_ConsoleStorageAPITests.js on fission. r=bomsy
The test was doing a navigation from about:blank to an example.com URL in a
ContentTask, which I guess was going bad with fission enabled since we were
creating a new content process.
I don't think there was a good reason to do that, so we start the test directly
with the "final" URL.
And since about:blank page can be special, we use a data url instead.
A few `info` calls are added to have better logs, and the test is slightly refactored
to be easier to follow.

Finally, the test was a bit loose, only checking the number of events we were
receiving. We take this opportunity to really check that we get the events that
are expected.

Differential Revision: https://phabricator.services.mozilla.com/D109278

UltraBlame original commit: aef5baafb94612ac5431bcabea9dd32541f7ab7f