Bug 1699736 - Update testcase for bug 1698543 to ignore stack space exhaustion in regexp compilation r=sfink
The problem here is that the test relies on triggering overrecursion but this can also trigger regexp compilation to fail with the error 'regexp too big'. It's ambiguous as to whether this indicates overrecursion or a regexp that's too large, so I don't think we can check for this in the test runner.

I checked that the testcase still triggers the original problem.

Differential Revision: https://phabricator.services.mozilla.com/D109292

UltraBlame original commit: 438ac889a6392ea938fa83f49749ac818846e4b0