Bug 1700151 - Set eglSwapInterval for each subsurface, r=jrmuizel
We need to set the swap interval with the egl surface being selected
via `eglMakeCurrent` - just setting it globally is not enough.

By default the value is `1`, which would cause the Wayland EGL platform
to request frame callbacks and block on `eglSwapBuffers`. This would
make the whole window block as soon as a single subsurface is obscurred.
Instead, we only want to block on the frame callback for the root surface.

Differential Revision: https://phabricator.services.mozilla.com/D109373

UltraBlame original commit: bc8866703de1077ac25f733155051b003143d768