Bug 1678892 - Pass dirty and opaque regions to layers compositor. r=sotaro
The backend CompositorD3D11/CompositorOGL layers compositors can already do
partial clear optimizations for us if applicable. The only thing we need to
do is pass in the actual dirty/opaque regions so that it can utilize it.

Like with RenderCompositorSWGL, we move the actual allocation of the framebuffer
into StartCompositing when this information is known, rather than BeginFrame
which is too early in the frame to have this information yet.

Differential Revision: https://phabricator.services.mozilla.com/D109092

UltraBlame original commit: 8f5b22203c79ce916167b2a33dd1d90f1d5f63ab