Bug 1697666 followup - Make the osaka-mono-exists test confirm that if Osaka is present, then Osaka-Mono is also available. r=arai
This version of the test passes for me locally with Osaka/Osaka-Mono installed, and also passes
(because both test and reference fall back to serif) if they're absent. It should fail only if
the Osaka family is present, but the Osaka-Mono family name (which is the special-case entry
that was broken) is not recognized.

Differential Revision: https://phabricator.services.mozilla.com/D109223

UltraBlame original commit: 14feeaed5794c069894a7584a191dd055e54beda