Bug 1700616: Specialize phis that only depend on OsrValues r=jandem
The new implementation of branch pruning in bug 1697696 can remove the path from the entry block to the OSR preheader. In this case, phis in the preheader only have one argument and can be removed, which means that phis elsewhere can have multiple OsrValues as operands.

In the original pruning patches, I handled the case where a phi *only* had OsrValue operands by forcing its type to `MIRType::Value`, but I didn't cover the case where we have a cycle of phis that only depend on each other and on OsrValue operands. This patch fixes that.

I initially tried adding a new MIRType to represent this case during phi specialization, but decided that changing `phi->type() == MIRType::None` to `phi->type() == MIRType::None || phi->type() == MIRType::OsrValue` everywhere in `TypeAnalyzer` wasn't actually an improvement.

Note that when we specialize these phis, we are *not* propagating `MIRType::Value` to their uses. In an earlier version of this patch I tried that, but realized it's unnecessary. If the use has a non-Value type, then `adjustPhiInputs` will add an optimistic unbox. (The testcase includes a case where we specialize incorrectly and bail out, to verify that this works.)

Differential Revision: https://phabricator.services.mozilla.com/D109686

UltraBlame original commit: e8691d90fb5733a39d23cf712e5673e4b761b813