Bug 1581859: Part 4a - Add skeleton PExtensions actor. r=zombie,nika
This is the start of an actor which will be automatically instantiated in the
parent and each content process which can be used to route most process-level
IPC traffic needed by the extensions framework. It should allow the extensions
framework to keep its IPC glue close to the code that uses it, and simplify
matters for child-side code which needs to run in both parent and content
processes.

Differential Revision: https://phabricator.services.mozilla.com/D103212

UltraBlame original commit: 2d4a14c67ba4726ab15ecc8110bf50e507445c28