Bug 1691579 - [devtools] Enable browser_spectrum.js on Fission. r=jdescottes.
The test failure doesn't seem to have anything to do with Fission,
but this might have introduce some changes in timings that would
trigger a race condition somehow.
The test was failing because in one of the test we were moving a slider
to change a color, while the slider element had a height of 0, which
would interfere with the computations that are done later.
This patch adds a helper function to create the Spectrum widget, wait for the
slider to have a non-zero height and then calling `show` (where some dimensions
are being cached).

Differential Revision: https://phabricator.services.mozilla.com/D109876

UltraBlame original commit: 5c8e536a9b1193667b9e6965530c8fb08154b90e