Bug 1701137: Properly parse taskcluster Sentry secret r=aki
Taskcluster secrets are not provided to clients in the same
format that they are set. Rather, the YAML is rendered
to JSON, and it's wrapped in a larger object that also
contains the `expires` timestamp.

Correctly parse the JSON and remove the dependency on `yq`.

Differential Revision: https://phabricator.services.mozilla.com/D109907

UltraBlame original commit: 6e52d2a6585df26518f5d6e9de03b2486e8f8f4c