Bug 1699224 - Part 1. Ensure we map/unmap surfaces used by DrawTargetSkia. r=jrmuizel
Since we want to be able to unmap shared surfaces which are not in use,
we need to be able to map the data in as required. Blob images can
access shared surfaces via recordings which are used by DrawTargetSkia.
This patch makes it call Map/Unmap for data surfaces that require a data
pointer in Skia.

Differential Revision: https://phabricator.services.mozilla.com/D109437

UltraBlame original commit: d2294fe6ef3120484d2b041ac6bd43e6c6815ff6