Bug 1689519 - Pre: Add AppUpdater.jsm helpers: terminal predicate, debug string. r=adw,bytesized
The AppUpdater module invokes a callback with updater status codes.
The callback can be invoked multiple times with different codes
throughout the lifecycle of an update (check).  This helper captures
the codes that are terminal, meaning that the update (check) has
completed.  It also includes a tiny helper to determine a string
representation of a status code for logging.

Differential Revision: https://phabricator.services.mozilla.com/D104637

UltraBlame original commit: cd5b5cc01886657fdc4a02cc81647bfb847b2043