Bug 1662559 - Part 5: Add .privateBrand to class body scopes. r=arai
Not every class gets this special binding. It will be needed for all classes
that have any nonstatic private methods, getters, or setters. With this patch,
we create a .privateBrand binding in the scopes for such classes, and populate
it with a symbol when initializing the class. However, it is not yet being used
anywhere; the constructor does not yet stamp instances with the private brand.

Depends on D108288

Differential Revision: https://phabricator.services.mozilla.com/D108289

UltraBlame original commit: e200366e7e3af667c0bd20acb41cd29413597da7