Bug 1700622 - Move --toolbar-field-non-lwt-bgcolor to browser-custom-colors.inc.css. r=ntim
This causes the search bar on the Customize page to fall back to Field, which will be themed correctly. Note that if the system is in Dark mode and Firefox is in light mode, the search bar will be dark. That should be fixed by bug 1700371.

Differential Revision: https://phabricator.services.mozilla.com/D110312

UltraBlame original commit: 6b5d91e2666144c336be0c175e64fd4bae1545a5