Bug 1691907 - Part 2: Add a method `initWithURI` to nsICookieJarSettings. r=dimi,robwu
This patch adds a method `initwithURI` to nsICookieJarSetting in order
to allow JS code to be able to set the paritionKey of the
CookieJarSettings.

This is needed for web extension for creating and setting the
cookieJarSettings for the download resource.

Differential Revision: https://phabricator.services.mozilla.com/D109759

UltraBlame original commit: 379fd9a10dd8e892dfd6a45c39ec9e429a9d3641