Bug 1701412: Use bookmark instead of visit for autofill r=eeejay
Previously we relied on adding a visit to prompt the URL bar to autofill, that worked because nightly and release shipped with bookmarks to addons.moz. Beta, however, does not, so we should modify this test to populate the bookmark on all releases.

Differential Revision: https://phabricator.services.mozilla.com/D110571

UltraBlame original commit: fc68f9ca86aada5668e5e3679d07ae4f029e1468