Bug 1701899 - Update object flags in a few places where properties are added. r=jonco
When freezing objects, we use a fast path where we didn't set the
HasNonWritableOrAccessorPropExclProto flag.

With the planned shape changes, we will hopefully be able to change this so that
we use a shared code path so that this kind of issue can't happen.

Differential Revision: https://phabricator.services.mozilla.com/D110246

UltraBlame original commit: e833a5f2ec97b1c48a9a8dc185ff68f4f6915d0f