Bug 1702662 - Fix scale_factors calculation in presence in negative ScaleOffset. r=gfx-reviewers,jrmuizel
The `scale_factors` method for `ScaleOffset` needs to take the absolute
value of the scale field, which is the same way that scale factors
are calculated in the slower transform based path. These scale factors
are used to determine the appropriate size of an offscreen surface in
some cases.

Differential Revision: https://phabricator.services.mozilla.com/D110726

UltraBlame original commit: 19b0de3b71064e1bcdcee18d550c99408e0b8ca1