Bug 1700850 - Pre: Create backgroundtask temporary profile before profile service. r=mossop
It's very unlikely that this would make a difference, since creating
the service does not immediately look for a profile.  But it's also
not clear to me why I originally placed it after the service was
created, and it's not impossible that the service creation would look
for `XRE_PROFILE_PATH` (which is how we communicate the desired
profile to the profile service), so let's create the profile before
the service.

Differential Revision: https://phabricator.services.mozilla.com/D110309

UltraBlame original commit: 3825a9d1a1e00b94915a814fced61135c2ee0782