Bug 1702245 - Expand XPCOM's in-tree documentation, r=xpcom-reviewers,kmag
This change involved migrating a number of pages, including the DMD
documentation, XPIDL documentation, and the Internal String Guide to
Firefox Source Docs, as well as performing a review/update to these
documents to be more accurate and current.

In addition, some additional sections were written in the various
documents to elaborate on important features which were not present when
the documents were originally written, and a few new documents were
written to add basic documentaiton for other XPCOM components.

The following MDN pages were migrated in this patch:

https://developer.mozilla.org/en-US/docs/Mozilla/Tech/XPIDL
https://developer.mozilla.org/en-US/docs/Mozilla/Tech/XPCOM/Guide/Internal_strings
https://developer.mozilla.org/en-US/docs/Mozilla/Performance/DMD

Differential Revision: https://phabricator.services.mozilla.com/D110424

UltraBlame original commit: ec7220b67affa44a91830b937050c3c3598852df