Bug 1702516 - Remove all terminateGlobal uses r=kmag
This never actually worked, because nothing ever listened to the notification
sent in XPCJSContext (kill-content-script-sandbox). We're also removing the
"Temporarily disable extension" button in the slow addon notification, which
was the primary caller of this - however, that's in a separate patch.

Differential Revision: https://phabricator.services.mozilla.com/D110583

UltraBlame original commit: fc396e06eb5f99ff6901e2f91989fd176d8f6725