Bug 1698321 - Make sure that we do not underflow mNumActiveConns r=necko-reviewers,kershaw
This may happen when a TransportSetup retry to connect using a different IP address, but AsyncResolveNative failed.

The change adds a flag that indicates whether it is needed to remove a connection from mNumActiveConns. This should make sure that we do not do it multiple times.

Differential Revision: https://phabricator.services.mozilla.com/D110723

UltraBlame original commit: 71aec1309f627ac29f58446884cf600d676eb4ed