Bug 1700076 - Add ability to cancel TRR requests before they're dispatched r=necko-reviewers,dragana
A TRR request is created and dispatched to the handling thread.
If we want to cancel it, we may do so before the runnable has been handled
and there's no HTTP channel - so the cancel was a NO-OP.
Here we add mCancelled which is consulted in SendHTTPRequest to check
if the request has already been cancelled.

Differential Revision: https://phabricator.services.mozilla.com/D109590

UltraBlame original commit: c085324cb20e25c2d68c00d48df2e55cac50c717