Bug 1675844 - comments, test cases, commit msg.  r=jseward
Commit message, comments, and test cases for s-s patch that landed
separately.

Several interlocking parts to the baseline part of the patch:

First, the pre-reservation of the value stack in the decode loop does
not work for opcodes with more than a fixed number of results: a
function or a block can return a larger number of values than any
reasonable pre-reserved size.  The pushResults method must reserve
space for the values it will push, and must propagate OOM up the call
tree.  This requires a fair amount of plumbing but is straightforward.

Second, the way the reservation should be made both in pushResults and
in the decode loop is to take into account the max number of pushes
per opcode, which can be defined as a constant (here very
conservatively as 10), but which is not "2" as asserted by a comment
in the existing code, and even if it were "2" it would not be that for
the reason given in the comment.

Third, the reservation code in the decode loop is cleaned up by
removing the guard that runs reservation code only once every 50
decode iterations, since this guard was always of questionable utility
and seems only to complicate the reasoning around the reservation
logic.

Finally, the guarded reservation code also contained a check for
available space in the temp allocator the baseline compiler uses for
allocating delayed code generation objects.  This check was necessary
because that allocator is infallible.  However, the compiler uses the
allocator rarely, and all existing uses of the allocator check for
allocation failure already and propagate the OOM, so the reservation
check can be removed if the allocator is simply marked as fallible.

On the ion side of things:

Every time we process an open-ended list of values for which we may
allocate nodes, we must call ensureBallast on the allocator, since it
is infallible.  This was done in one place in WasmIonCompile, but not
both places where it was needed.

As far as the test case is concerned:

This test was enough to find the errors in both baseline and ion.  It
can probably be pared down from the 100..300 loop it has now to
something smaller, but there you have it.

Differential Revision: https://phabricator.services.mozilla.com/D98513

UltraBlame original commit: e23c3e39ac59190cb7d4afa84dbf0333e7d167d3