Bug 1658914 - Stop sampling animations via vsync ticks on WebRender when we are in testing refresh driver mode. r=boris
On WebRender there are two functions to be called for sampling animations, one is
OMTA::Sample which is called via vsync ticks in normal mode, the other is
OMTA::SampleForTesting which is called in testing refresh driver mode. The
former function is called on the sampler thread, if it gets called in between
we call nsIDOMWindowUtils.advanceTimeAndRefresh and
nsIDOMWindowUtils.getOMTAStyle AND IF the result of sampling via getOMTAStyle
was `SampleResult::None` or `SampleResult::Skipped`, we will use an incorrect
animation value via the OMTA::Sample call with the vsync timestamp, the vsync
timestamp is completely different from the value advanced by
the advanceTimeAndRefresh call.

Differential Revision: https://phabricator.services.mozilla.com/D110888

UltraBlame original commit: ef63364742306d3fb8d12a8ebe4a827ba576409d