Bug 1703472 - For window.focus, popup abuse level should be checked on the caller. r=smaug,hsivonen
Just like how for window.open() it's checked on the opener, of course.

Note that when sending the focus request via IPC, we trust the check
done on the child (bug 1677899).

Do we have a good test to extend to cover this case?

Differential Revision: https://phabricator.services.mozilla.com/D111107

UltraBlame original commit: db92ea730300b6e9f4dd49075437e866190ee9e3