Bug 1704042 part 9 - Report an exception for large lengths from JS::GetArrayLength API. r=lth
There are a number of callers in Gecko, most of them are using this to iterate over an object
expected to be an array. This should be okay for now; if there are callers that need the
full length we can add a separate API that returns uint64_t.

This lets us remove ToLengthClamped and the GetLengthProperty overload returning uint32_t.

Differential Revision: https://phabricator.services.mozilla.com/D111389

UltraBlame original commit: 79f3f364ed334760f3026af08b9611e74aa62d3c