Bug 1682195 - Use scissoring to restrict cs_clip_image rather than discard. r=gw
cs_clip_image renders the entire target-space sub-rect, and then uses discard against
the local-space bounds to ensure the primitive gets trimmed down to its actual footprint
within the larger target-space sub-rect. This can be fairly wasteful and slow.

This patch restructures things so that we essentially render with the actual primitive
vertexes, and then we either clamp (for axis-aligned) or scissor (for non-axis-aligned)
to restrict the footprint to within the target-space sub-rect.

This then allows the use of a SWGL fast-path for the axis-aligned case that can vastly
speed up the common-case.

Differential Revision: https://phabricator.services.mozilla.com/D111316

UltraBlame original commit: b641b7f90108a2e6b57186139b450d9ccb80c96c