Bug 1702196 - Add testing functions for verifying shape invariants. r=jonco
This allows taking a snapshot of an object's state and then at a later point
checking invariants that we expect to hold for shapes. This will let us catch
bugs where an object or shape is mutated in ways that would be a problem for
JIT code.

Differential Revision: https://phabricator.services.mozilla.com/D111266

UltraBlame original commit: e1ed88d29409462d3d147418dffd02f61e460618