Bug 872078 - Added preview and info for fonts in Response view of Netmonitor. r=bomsy
A preview for fonts is added to the Response view of the Netmonitor. This view displays the name and MIME type of the font and generates a preview for it.

If no preview can be created, a hint is shown to the user. This is the case when the font isn't used within the page when generating the preview. For example, this is the case when loading the font via the Font Loading API.

Differential Revision: https://phabricator.services.mozilla.com/D110169

UltraBlame original commit: f73193dcd637fa5722acb727e396caa350346138