Bug 1701749. When computing the double tap rect if there is a overflow rect that contains the bounding client rect use that. r=emilio
I'm not sure if this is better overall or not, it's better for the testcase in this bug (a div with a large amount of visible overflow).

This code is a direct port of javascript code, so it only had the option of using getBoundingClientRect.

Differential Revision: https://phabricator.services.mozilla.com/D110534

UltraBlame original commit: e0df81346b2baf15aab5698106bfb0509d2e7642