Bug 1704141 - Ensure bytecode is deduplicated when decoding StencilXDR. r=nbp
While we don't leak memory, this is a missed memory savings and also results
in the data showing up as heap-unclassified. The legacy-XDR code does this
sharing, but the feature was broken in Stencil-XDR and this patch fixed it.

Differential Revision: https://phabricator.services.mozilla.com/D111452

UltraBlame original commit: 537141f364c56c576b453b848d6be8104c0f5003