Bug 1704545, part 1 - Make the priority manager's OnRemoteBrowserFrameShown into a direct call. r=gsvelto
ParticularProcessPriorityManager observes remote-browser-shown and recomputes
the priority for the given browser. This patch eliminates the observer call
and instead makes the frame loader do a direct call.

I think doing it this way is a little clearer, and also the next patch will
eliminate the other reason ParticularProcessPriorityManager is an observer,
which will let us not not register something as an observer for every content
process, which I think is also good.

There should be no change in behavior.

Differential Revision: https://phabricator.services.mozilla.com/D111653

UltraBlame original commit: 3c070aae7b48737bd4f5fec7db243efdd8bf58ea