Bug 1698843 - Part 2: Make the private cookieBehavior getter to return the regular pref if there is a user value for the regular pref. r=johannh,dimi
If the user has a non-default regular cookieBehavior, we will make the
private cookieBehavior getter to return the regular pref to mirror the
cookieBehavior in ETP custom mode.

In addition, we don't need to do the pref migration because if the user
has a non-default cookieBehavior, it will directly mirror to the private
cookieBehavior pref so that the cookieBehavior is consistent for private
mode.

Differential Revision: https://phabricator.services.mozilla.com/D111002

UltraBlame original commit: 4c5fb9d17322f92c1a0d1f1681541aa7473e25f9