Bug 1704034 - Make the stackwalker print out the name of NTSTATUS error codes r=KrisWright
This patch changes a few things regarding how error codes are printed out.
First of all it updates the list of statuses defined in ntstatus.h, and makes
it up-to-date with the current Windows SDK. With this done we change how we
print out a minidump exception code so that if it's not among the known
EXCEPTION_* values we try to print it out as a STATUS_* one.

It's worth noting that the EXCEPTION_* definitions are just aliaeses for
STATUS_* ones and deprecated in Windows but since they're commonly used we
just keep them.

Last but not least we change how EXCEPTION_BOUNDS_EXCEEDED,
EXCEPTION_STACK_BUFFER_OVERRUN and EXCEPTION_HEAP_CORRUPTION are
printed because they are inconsistent with their Windows equivalents.
They're "breakpadisms" and they've caused confusion in the past.

Differential Revision: https://phabricator.services.mozilla.com/D111712

UltraBlame original commit: d547d751a901e55583625dc7d298aa4930e8c3f0