Bug 1702466 - Avoid a lot of IPC traffic and extra content state changes when mousing over menulist items. r=Gijs
We're sending two messages (one out, one in) when we switch between
menuitems for example. The patch by Matt caused us to sometimes paint in
between that period of time and made this noticeable, but it seems
pretty bad regardless.

Send messages only when the mouse enters / leaves the popup, which is
what the code was intended to do.

Differential Revision: https://phabricator.services.mozilla.com/D112035

UltraBlame original commit: 2a778a05368b539ccae766ba5ea528ec7073cfd6