Bug 1700215 - Make OverscrollAnimation coexist with momentum scrolling in the other axis. r=botond
This change makes us match Chrome's overscroll behavior (i.e. restoring
overscroll state on an axis while scrolling by pan momentum keeps going on the
other axis).

Basic mechanisms to do the behavior are;

a) if there is an on-going overscroll animation on an axis and the overscroll
   direction is opposed to the momentum direction, stop the overscroll animation
   immediately and keep the momentum scroll (or overscroll and scroll if the
   current position is in overscroll gutter).
b) if there is an on-going overscroll animation on an axis and the overscroll
   direction is same as the momentum direction, drop the momentum and keep
   the animation overscrolling.
c) if there is an on-going overscroll animation on an axis and no overscroll
   animation on the other axis and if a pan momentum event overscrolls on
   the other axis,  start a new overscroll animation on the axis.

With this change there still remains a (relatively) less noticeable issue that
pan momentum scrolling on an axis stops immediately after the other axis
overscroll animation stopped.  For example, given that there is a horizontal
overscroll animation and vertical pan momentums keep going, subsequent vertical
pan moments after the horizontal overscroll animation finished will be discarded
since we do set the APZC state to NOTHING in
AsyncPanZoomController::UpdateAnimation(). To fix this issue we need to handle
OVERSCROLL_ANIMATION separately from the others, in the example case, once after
the overscroll animation finished, the state should be MOMENTUMPAN, thus we can
keep consuming momentum events as scroll. That said, the fix will not be easy,
because there are also cumbersome cases where we don't want to keep consuming
the events, one of the cases is b) in the basic mechanisms above, in the case
we do want to discard incoming momentum events after the overscroll animation
finished if the incoming momentum events are in the initial overscroll gutter,
IF the incoming momentum events are in the another edge of the overscroll
gutter, we have to re-initiate a new overscroll animation in the other side
gutter. We will handle these cases in a followup bug (bug 1702949).

Differential Revision: https://phabricator.services.mozilla.com/D110794

UltraBlame original commit: 15023df770afce9a82e81fc2a75031ded5f2d2cf