Bug 1703318 - Part 3: Add Marionette "background-update" ping test. r=chutten,bytesized
This is subtle but not so subtle that it can't be understood.  The
major issue we might see is that the background update tasks all use
the same `UpdRootD` and `app.update.background.enabled` state,
essentially preventing concurrent Marionette tests.  In the future we
may be able to do better.

Differential Revision: https://phabricator.services.mozilla.com/D111349

UltraBlame original commit: eef49d5b41e5119ae0d697b61740f541ee411b51