Bug 1704071 - Increase address bar and search bar contrast during toolbar customization. r=harry
To improve the visibility of these input fields this pathc adds a border based
on currentColor and reuse the urlbar shadow. This is a small improvement over the
status-quo, the applied transparencies are controlled by the drag'n'drop code.
This issue exists also in pre-proton, so the styles apply regardless.

Differential Revision: https://phabricator.services.mozilla.com/D112030

UltraBlame original commit: e76fec22745d6dcd3ce9cd8a79927841e8ed30ef