Bug 1705599. Allow double tap to zoom rect calculation code to work on fixed content. r=botond
nsLayoutUtils::GetBoundingFrameRect has one minor issue with the scroll clip, if we are in a scroll frame that inside fixed content it won't be a descendant of the root scroll frame.

So we switch to the more generic nsLayoutUtils::TransformRect instead of TransformFrameRectToAncestor.

Differential Revision: https://phabricator.services.mozilla.com/D112306

UltraBlame original commit: b998b017136451654a4472fc087dba9cc29ae69a