Bug 1705622. When building the hit testing tree use async zoom containers to keep track of which zoom constraints to apply. r=botond
This also fixes a bug where root content xul docs had zoom constraints that allowed them to be zoomed. This would cause a crash because we would ask for a repaint request and it would get routed to APZCCallbackHelper::UpdateRootFrame (because it has the id of the document element), which assumes the existance of a root scroll frame.

Differential Revision: https://phabricator.services.mozilla.com/D112486

UltraBlame original commit: 788613ebb6894c59af33e9f6057b09acdbf4b256