Bug 1667235, fix various issues with the browser_selectpopup.js test, mostly on Windows, r=Gijs
Change the height of a window opened during the test now that menus on windows 10 are larger. Make sure that the zoom level is reset after the test so that running the test again doesn't use the old zoom level. The default value of 'ignorekeys' is 'shortcuts' on Windows so reset this properly otherwise the test fails if run again.

Differential Revision: https://phabricator.services.mozilla.com/D112407

UltraBlame original commit: b536dcb830ff04901b25ac6ac273d0b5184265c8