Bug 1691034: Remove problematic "auto-reload" behavior from crashtest 370174-3.html. r=emilio
This "automatic reload" behavior was probably useful as a PoC for the original
bug, but it's no good in a crashtest, because:

(a) the reload probably doesn't even take effect most of the time, since it's
in a setTimeout that the test is under no obligation to wait for.

(b) if the reload *does* get a chance to take effect (e.g. on a slower
test-runner where it takes longer for the test harness to judge the test to be
complete), then the test may simply never terminate (or take a long time to
terminate), because each reload causes some more paints, which the test harness
is then forced to wait on; and another reload may arrive before the test
harness checks back to see if we're done painting.

So: let's just remove the auto-reload behavior from this crashtest.

Differential Revision: https://phabricator.services.mozilla.com/D112686

UltraBlame original commit: eda6b244848e3dc4240ef0077178f0afab4fa4ce