Bug 1706573 - Don't generate a new shape for updated dictionary properties. r=jonco
When changing a dictionary property, we have to generate a new shape for the
object (last property shape) and we can then mutate the property in place.
This removes the extra shape allocation for the mutated property.

This reverts the change in bug 703157: the issue fixed there no longer applies.

`replaceWithNewEquivalentShape` is then only used for `generateOwnShape`, to generate
a new shape for the last property, so we can inline it into `generateOwnShape` and
simplify the code a bit.

Differential Revision: https://phabricator.services.mozilla.com/D112907

UltraBlame original commit: 8b8067f49855aef2343699ec5ab2a95efa141220