Bug 1553399 - Unlock nsRemoteService before launching child r=mossop
This fixes the issue for me. I can't think of any problems with doing this
here? However I am not an expert in the remote service. As part of this I also
cover the case where the user encounters the profile lock dialogue and selects
to kill the existing instance of Firefox. This can result in a slow startup in
a similar way as far as I've been able to observe.

Differential Revision: https://phabricator.services.mozilla.com/D113054

UltraBlame original commit: 53c5b8e49f491611b102b271ae14199dd94a2c3a