Bug 1707820 - Fix pre-barrier verifier assertion failure when adding dictionary properties. r=jonco
We were adding the new shape to the table first, but this confused the pre-barrier
verifier. Likely a false positive because the shape is newly allocated.

We now add the entry to the table at the end, and if this fails we remove the
dictionary shape again. This also removes some unnecessary code for growing slots.

Differential Revision: https://phabricator.services.mozilla.com/D113493

UltraBlame original commit: 9a0b198c9a5ddb582855f7a03c486938992a4057