Bug 1707823 - [devtools] Make navigateTo more secure. r=ochameau.
In the `BrowserTestUtils.browserLoaded` call we use in `navigateTo`, we weren't
checking for a specific uri to be loaded, which means the Promise could resolve
on *any* page being loaded.
This patch makes it more secure by waiting the load event for the uri we navigated to.

Differential Revision: https://phabricator.services.mozilla.com/D113479

UltraBlame original commit: 015d6eaef2cf61aa13479412d5e327dfae78c70e