Bug 1706868. Don't request content paint at the "after" resolution when starting a zoom animation. r=botond
This is a waste because we don't render that content, and if we did (we do sometimes, and it leads to flashing) it would be low resolution in the case of zooming out. It also causes bugs like bug 1698537. And if we get some other updates in there we can get stuck in a state where we half think we are zoomed in and half think we are zoomed out.

See bug 1698537, comment 11 for full back story.

Differential Revision: https://phabricator.services.mozilla.com/D113074

UltraBlame original commit: 0f75bd775447acb5c8c571ad7f1fe2206d1061a3