Bug 1706694 - Don't overwrite leftmost rope's internals at the start when reusing the leftmost child buffer in string flattening r=jandem
This is unnecessary and was causing problems. We can leave the leftmost rope to
be handled in the same way as all the other ropes, and move the special case to
just avoid copying the chars when reusing the leftmost buffer.

The problem was that now we always barrier ropes, before overwriting the left
child pointer with the parent pointer. If this happens during incremental GC we
must ensure we haven't already overwritten this pointer we the buffer pointer.

This patch also moves changing the leftmost string into a dependent string
until the end.

Depends on D113316

Differential Revision: https://phabricator.services.mozilla.com/D113503

UltraBlame original commit: 1d6f87631df729ae076d9d8ce5a4ae7f54491340