Bug 1704129 - Add a test to measure process switch time. r=sefeng,perftest-reviewers,Bebe
This patch adds a test to measure fission process switch time. It uses two recordings replayed at the same time, along with a custom browsertime pageload script to do this. Two urls are provided through the manifest and the second url is the one that gets measured.

* * *
Bug 1704129 - Add test_script field to raptor manifests. r?#perftest

This patch adds a `test_script` field to the raptor manifests that allows us to customize the browsertime test script to use. However, we only allow it to be used on pageload tests for the time being.

Differential Revision: https://phabricator.services.mozilla.com/D113220

UltraBlame original commit: 2801837a06b22ffc82605ea68c0e6382efdc62c5